Refactor sample
Refactor the sample into a better organization.
