+++ docs/csharp/language-reference/builtin-types/snippets/unions/Program.cs
@@ -1,0 +1,10 @@
+BasicUnionScenario.Run();
+GenericUnionScenario.Run();
+ValueTypeCasesScenario.Run();
+BodyMembersScenario.Run();
+NullHandlingScenario.Run();
+ManualUnionScenario.Run();
+NonBoxingAccessScenario.Run();
+ClassUnionScenario.Run();
+// Uncomment when union member providers are available in the compiler:
+// MemberProviderScenario.Run();
