+++ docs/csharp/language-reference/builtin-types/snippets/unions/RuntimePolyfill.cs
@@ -1,0 +1,11 @@
+// Remove this file when UnionAttribute and IUnion are included in the .NET runtime.
+namespace System.Runtime.CompilerServices
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = false)]
+    public class UnionAttribute : Attribute;
+
+    public interface IUnion
+    {
+        object? Value { get; }
+    }
+}
